No need for ObjectUtils, we can move remaining code into GarnetObjectBase. Allow long cursor for object SCAN commands.
